+++ resources/forms-ui/src/app/pages/dashboard/admin/admin-create-form-page/admin-create-form-page.component.ts
@@ -67,6 +67,8 @@ export class AdminCreateFormPageComponent implements OnInit {
             inputSets: form_elements,
             scrollToFieldOnAdd: false,
             disabledActionButtons: ['data', 'clear', 'save'],
+            typeUserAttrs: this.formBuilderService.handleFieldsTypeAttrs(),
+            typeUserDisabledAttrs: this.formBuilderService.disableFieldAttrs(),
             disableFields: this.formBuilderService.disableDefaultFormControls()
           });
 
@@ -88,6 +90,8 @@ export class AdminCreateFormPageComponent implements OnInit {
             scrollToFieldOnAdd: false,
             defaultFields: this.template.form_fields,
             disabledActionButtons: ['data', 'clear', 'save'],
+            typeUserAttrs: this.formBuilderService.handleFieldsTypeAttrs(),
+            typeUserDisabledAttrs: this.formBuilderService.disableFieldAttrs(),
             disableFields: this.formBuilderService.disableDefaultFormControls()
           });