+++ resources/forms-ui/src/app/pages/dashboard/form-templates/add-template-page/add-template-page.component.ts
@@ -42,6 +42,8 @@ export class AddTemplatePageComponent implements OnInit {
           inputSets: form_elements,
           scrollToFieldOnAdd: false,
           disabledActionButtons: ['data', 'clear', 'save'],
+          typeUserAttrs: this.formBuilderService.handleFieldsTypeAttrs(),
+          typeUserDisabledAttrs: this.formBuilderService.disableFieldAttrs(),
           disableFields: this.formBuilderService.disableDefaultFormControls()
         });