+++ resources/forms-ui/src/app/pages/dashboard/front-desk/front-desk-submitted-forms-list-page/front-desk-submitted-forms-list-page.component.ts
@@ -56,7 +56,7 @@ export class FrontDeskSubmittedFormsListPageComponent implements OnInit {
 
   open(e: Event, form: any) {
     e.stopPropagation();
-    this.router.navigateByUrl('/front_desk/preview', { state: { form: form }});
+    this.router.navigateByUrl('/front_desk/view_form', { state: { form: form }});
   }
 
   print(ev: Event, form: any) {