+++ resources/forms-ui/src/app/pages/welcome-page/welcome-page.component.ts
@@ -7,7 +7,11 @@ import { Component, OnInit } from '@angular/core';
 })
 export class WelcomePageComponent implements OnInit {
 
-  constructor() { }
+  year: string;
+
+  constructor() {
+    this.year = new Date().getFullYear().toString();
+  }
 
   ngOnInit() {
   }