+++ resources/forms-ui/src/app/services/form-builder/form-builder.service.ts
@@ -1,6 +1,5 @@
 import * as _ from 'lodash';
 import { Injectable } from '@angular/core';
-import { Forms } from 'src/app/models/forms.model';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { EndpointService } from '../endpoint/endpoint.service';
 import { SectionsService } from '../sections/sections.service';
@@ -11,11 +10,17 @@ import { SectionsService } from '../sections/sections.service';
 export class FormBuilderService {
 
   headers: HttpHeaders;
+  disableClassName: any;
   private formFieldClassName: string;
 
   constructor(private http: HttpClient, private endpointService: EndpointService, private sectionService: SectionsService) {
     this.headers = this.endpointService.headers();
     this.formFieldClassName = 'form-control';
+    this.disableClassName = {
+      className: {
+          style: 'pointer-events: none'
+        }
+    };
   }
 
   disableDefaultFormControls() {
@@ -25,6 +30,162 @@ export class FormBuilderService {
       'hidden',
       'number',
     ];
+  }
+
+  handleFieldsTypeAttrs() {
+    return {
+      text: this.disableClassName,
+      file: this.disableClassName,
+      button: this.disableClassName,
+      hidden: this.disableClassName,
+      number: this.disableClassName,
+      textarea: this.disableClassName,
+      autocomplete: this.disableClassName,
+    };
+  }
+
+  disableFieldAttrs() {
+    return {
+      'header': [
+        'access',
+        'className'
+      ],
+      'autocomplete': [
+        'access',
+        'className',
+        'placeholder',
+      ],
+      'button': [
+        'style',
+        'value',
+        'access',
+        'subtype',
+        'className'
+      ],
+      'checkbox-group': [
+        'other',
+        'toggle',
+        'access',
+        'className'
+      ],
+      'file': [
+        'access',
+        'subtype',
+        'className',
+        'placeholder',
+      ],
+      'hidden': [
+      ],
+      'number': [
+        'value',
+        'access',
+        'className',
+        'placeholder'
+      ],
+      'radio-group': [
+        'other',
+        'toggle',
+        'access',
+        'className'
+      ],
+      'select': [
+        'access',
+        'className',
+        'placeholder'
+      ],
+      'paragraph': [
+        'access',
+        'subtype',
+        'className'
+      ],
+      'textarea': [
+        'value',
+        'access',
+        'subtype',
+        'className',
+        'placeholder'
+      ],
+      'text': [
+        'value',
+        'access',
+        'subtype',
+        'className',
+        'placeholder'
+      ]
+    };
+  }
+
+  disableAllFieldAttrs() {
+    return {
+      'header': [
+        'access',
+        'className'
+      ],
+      'autocomplete': [
+        'access',
+        'className',
+        'placeholder',
+      ],
+      'button': [
+        'style',
+        'value',
+        'access',
+        'subtype',
+        'className'
+      ],
+      'checkbox-group': [
+        'other',
+        'toggle',
+        'access',
+        'className'
+      ],
+      'file': [
+        'access',
+        'subtype',
+        'className',
+        'placeholder',
+      ],
+      'hidden': [
+      ],
+      'number': [
+        'value',
+        'access',
+        'className',
+        'placeholder'
+      ],
+      'radio-group': [
+        'other',
+        'toggle',
+        'access',
+        'className'
+      ],
+      'select': [
+        'access',
+        'className',
+        'placeholder'
+      ],
+      'paragraph': [
+        'access',
+        'subtype',
+        'className'
+      ],
+      'textarea': [
+        'name',
+        'value',
+        'access',
+        'subtype',
+        'className',
+        'placeholder'
+      ],
+      'text': [
+        'name',
+        'value',
+        'access',
+        'subtype',
+        'className',
+        'placeholder'
+      ]
+    };
   }
 
   disableSectionFormFields() {